testEnv in the dedicated common_test.go
